Remove SubRepoPerms from non-enterprise entirely
