+++ cmd/frontend/graphqlbackend/user_emails_test.go
@@ -232,9 +232,6 @@ func TestSetUserEmailVerified(t *testing.T) {
 			userExternalAccounts := database.NewMockUserExternalAccountsStore()
 			userExternalAccounts.DeleteFunc.SetDefaultReturn(nil)
 
-			subrepoPerms := database.NewMockSubRepoPermsStore()
-			subrepoPerms.DeleteByUserFunc.SetDefaultReturn(nil)
-
 			db := database.NewMockDB()
 			db.TransactFunc.SetDefaultReturn(db, nil)
 			db.DoneFunc.SetDefaultHook(func(err error) error {
@@ -245,7 +242,6 @@ func TestSetUserEmailVerified(t *testing.T) {
 			db.UserEmailsFunc.SetDefaultReturn(userEmails)
 			db.AuthzFunc.SetDefaultReturn(authz)
 			db.UserExternalAccountsFunc.SetDefaultReturn(userExternalAccounts)
-			db.SubRepoPermsFunc.SetDefaultReturn(subrepoPerms)
 
 			RunTests(t, test.gqlTests(db))
