+++ enterprise/cmd/gitserver/shared/shared.go
@@ -16,8 +16,7 @@ import (
 func EnterpriseInit(db database.DB) {
 	logger := log.Scoped("enterprise", "gitserver enterprise edition")
 	var err error
-	database.SubRepoPermsWith = edb.SubRepoPermsWith
-	authz.DefaultSubRepoPermsChecker, err = srp.NewSubRepoPermsClient(db.SubRepoPerms())
+	authz.DefaultSubRepoPermsChecker, err = srp.NewSubRepoPermsClient(edb.NewEnterpriseDB(db).SubRepoPerms())
 	if err != nil {
 		logger.Fatal("Failed to create sub-repo client", log.Error(err))
 	}
