+++ enterprise/internal/database/sub_repo_perms_store.go
@@ -48,12 +48,15 @@ type SubRepoPermsStore interface {
 	DeleteByUser(ctx context.Context, userID int32) error
 }
 
+
 // subRepoPermsStore is the unified interface for managing sub repository
 // permissions explicitly in the database. It is concurrency-safe and maintains
 // data consistency over sub_repo_permissions table.
 type subRepoPermsStore struct {
 	*basestore.Store
 }
+
+var _ SubRepoPermsStore = (*subRepoPermsStore)(nil)
 
 func SubRepoPermsWith(other basestore.ShareableStore) SubRepoPermsStore {
 	return &subRepoPermsStore{Store: basestore.NewWithHandle(other.Handle())}
