+++ enterprise/internal/gitserver/integration_tests/tree_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	edb "github.com/sourcegraph/sourcegraph/enterprise/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	inttests "github.com/sourcegraph/sourcegraph/internal/gitserver/integration_tests"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -49,7 +50,7 @@ func TestReadDir_SubRepoFiltering(t *testing.T) {
 		},
 	})
 	defer conf.Mock(nil)
-	srpGetter := database.NewMockSubRepoPermsStore()
+	srpGetter := edb.NewMockSubRepoPermsStore()
 	testSubRepoPerms := map[api.RepoName]authz.SubRepoPermissions{
 		repo: {
 			Paths: []string{"/**", "-/app/**"},
