+++ internal/database/database.go
@@ -45,7 +45,6 @@ type DB interface {
 	SavedSearches() SavedSearchStore
 	SearchContexts() SearchContextsStore
 	Settings() SettingsStore
-	SubRepoPerms() SubRepoPermsStore
 	TemporarySettings() TemporarySettingsStore
 	UserCredentials(encryption.Key) UserCredentialsStore
 	UserEmails() UserEmailsStore
@@ -203,10 +202,6 @@ func (d *db) SearchContexts() SearchContextsStore {
 
 func (d *db) Settings() SettingsStore {
 	return SettingsWith(d.Store)
-}
-
-func (d *db) SubRepoPerms() SubRepoPermsStore {
-	return SubRepoPermsWith(d.Store)
 }
 
 func (d *db) TemporarySettings() TemporarySettingsStore {
