+++ internal/database/mocks_temp.go
@@ -15,7 +15,6 @@ import (
 	uuid "github.com/google/uuid"
 	sqlf "github.com/keegancsmith/sqlf"
 	api "github.com/sourcegraph/sourcegraph/internal/api"
-	authz "github.com/sourcegraph/sourcegraph/internal/authz"
 	conf "github.com/sourcegraph/sourcegraph/internal/conf"
 	basestore "github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	encryption "github.com/sourcegraph/sourcegraph/internal/encryption"
@@ -3722,9 +3721,6 @@ type MockDB struct {
 	// SettingsFunc is an instance of a mock function object controlling the
 	// behavior of the method Settings.
 	SettingsFunc *DBSettingsFunc
-	// SubRepoPermsFunc is an instance of a mock function object controlling
-	// the behavior of the method SubRepoPerms.
-	SubRepoPermsFunc *DBSubRepoPermsFunc
 	// TemporarySettingsFunc is an instance of a mock function object
 	// controlling the behavior of the method TemporarySettings.
 	TemporarySettingsFunc *DBTemporarySettingsFunc
@@ -3928,11 +3924,6 @@ func NewMockDB() *MockDB {
 		},
 		SettingsFunc: &DBSettingsFunc{
 			defaultHook: func() (r0 SettingsStore) {
-				return
-			},
-		},
-		SubRepoPermsFunc: &DBSubRepoPermsFunc{
-			defaultHook: func() (r0 SubRepoPermsStore) {
 				return
 			},
 		},
@@ -4163,11 +4154,6 @@ func NewStrictMockDB() *MockDB {
 				panic("unexpected invocation of MockDB.Settings")
 			},
 		},
-		SubRepoPermsFunc: &DBSubRepoPermsFunc{
-			defaultHook: func() SubRepoPermsStore {
-				panic("unexpected invocation of MockDB.SubRepoPerms")
-			},
-		},
 		TemporarySettingsFunc: &DBTemporarySettingsFunc{
 			defaultHook: func() TemporarySettingsStore {
 				panic("unexpected invocation of MockDB.TemporarySettings")
@@ -4326,9 +4312,6 @@ func NewMockDBFrom(i DB) *MockDB {
 		},
 		SettingsFunc: &DBSettingsFunc{
 			defaultHook: i.Settings,
-		},
-		SubRepoPermsFunc: &DBSubRepoPermsFunc{
-			defaultHook: i.SubRepoPerms,
 		},
 		TemporarySettingsFunc: &DBTemporarySettingsFunc{
 			defaultHook: i.TemporarySettings,
@@ -7766,104 +7749,6 @@ func (c DBSettingsFuncCall) Args() []interface{} {
 // Results returns an interface slice containing the results of this
 // invocation.
 func (c DBSettingsFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// DBSubRepoPermsFunc describes the behavior when the SubRepoPerms method of
-// the parent MockDB instance is invoked.
-type DBSubRepoPermsFunc struct {
-	defaultHook func() SubRepoPermsStore
-	hooks       []func() SubRepoPermsStore
-	history     []DBSubRepoPermsFuncCall
-	mutex       sync.Mutex
-}
-
-// SubRepoPerms delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockDB) SubRepoPerms() SubRepoPermsStore {
-	r0 := m.SubRepoPermsFunc.nextHook()()
-	m.SubRepoPermsFunc.appendCall(DBSubRepoPermsFuncCall{r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the SubRepoPerms method
-// of the parent MockDB instance is invoked and the hook queue is empty.
-func (f *DBSubRepoPermsFunc) SetDefaultHook(hook func() SubRepoPermsStore) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// SubRepoPerms method of the parent MockDB instance invokes the hook at the
-// front of the queue and discards it. After the queue is empty, the default
-// hook function is invoked for any future action.
-func (f *DBSubRepoPermsFunc) PushHook(hook func() SubRepoPermsStore) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *DBSubRepoPermsFunc) SetDefaultReturn(r0 SubRepoPermsStore) {
-	f.SetDefaultHook(func() SubRepoPermsStore {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *DBSubRepoPermsFunc) PushReturn(r0 SubRepoPermsStore) {
-	f.PushHook(func() SubRepoPermsStore {
-		return r0
-	})
-}
-
-func (f *DBSubRepoPermsFunc) nextHook() func() SubRepoPermsStore {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *DBSubRepoPermsFunc) appendCall(r0 DBSubRepoPermsFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of DBSubRepoPermsFuncCall objects describing
-// the invocations of this function.
-func (f *DBSubRepoPermsFunc) History() []DBSubRepoPermsFuncCall {
-	f.mutex.Lock()
-	history := make([]DBSubRepoPermsFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// DBSubRepoPermsFuncCall is an object that describes an invocation of
-// method SubRepoPerms on an instance of MockDB.
-type DBSubRepoPermsFuncCall struct {
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 SubRepoPermsStore
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c DBSubRepoPermsFuncCall) Args() []interface{} {
-	return []interface{}{}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c DBSubRepoPermsFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
 }
 
@@ -40209,1527 +40094,6 @@ func (c SettingsStoreWithFuncCall) Args() []interface{} {
 // Results returns an interface slice containing the results of this
 // invocation.
 func (c SettingsStoreWithFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// MockSubRepoPermsStore is a mock implementation of the SubRepoPermsStore
-// interface (from the package
-// github.com/sourcegraph/sourcegraph/internal/database) used for unit
-// testing.
-type MockSubRepoPermsStore struct {
-	// DeleteByUserFunc is an instance of a mock function object controlling
-	// the behavior of the method DeleteByUser.
-	DeleteByUserFunc *SubRepoPermsStoreDeleteByUserFunc
-	// DoneFunc is an instance of a mock function object controlling the
-	// behavior of the method Done.
-	DoneFunc *SubRepoPermsStoreDoneFunc
-	// GetFunc is an instance of a mock function object controlling the
-	// behavior of the method Get.
-	GetFunc *SubRepoPermsStoreGetFunc
-	// GetByUserFunc is an instance of a mock function object controlling
-	// the behavior of the method GetByUser.
-	GetByUserFunc *SubRepoPermsStoreGetByUserFunc
-	// GetByUserAndServiceFunc is an instance of a mock function object
-	// controlling the behavior of the method GetByUserAndService.
-	GetByUserAndServiceFunc *SubRepoPermsStoreGetByUserAndServiceFunc
-	// HandleFunc is an instance of a mock function object controlling the
-	// behavior of the method Handle.
-	HandleFunc *SubRepoPermsStoreHandleFunc
-	// RepoIDSupportedFunc is an instance of a mock function object
-	// controlling the behavior of the method RepoIDSupported.
-	RepoIDSupportedFunc *SubRepoPermsStoreRepoIDSupportedFunc
-	// RepoSupportedFunc is an instance of a mock function object
-	// controlling the behavior of the method RepoSupported.
-	RepoSupportedFunc *SubRepoPermsStoreRepoSupportedFunc
-	// TransactFunc is an instance of a mock function object controlling the
-	// behavior of the method Transact.
-	TransactFunc *SubRepoPermsStoreTransactFunc
-	// UpsertFunc is an instance of a mock function object controlling the
-	// behavior of the method Upsert.
-	UpsertFunc *SubRepoPermsStoreUpsertFunc
-	// UpsertWithSpecFunc is an instance of a mock function object
-	// controlling the behavior of the method UpsertWithSpec.
-	UpsertWithSpecFunc *SubRepoPermsStoreUpsertWithSpecFunc
-	// WithFunc is an instance of a mock function object controlling the
-	// behavior of the method With.
-	WithFunc *SubRepoPermsStoreWithFunc
-}
-
-// NewMockSubRepoPermsStore creates a new mock of the SubRepoPermsStore
-// interface. All methods return zero values for all results, unless
-// overwritten.
-func NewMockSubRepoPermsStore() *MockSubRepoPermsStore {
-	return &MockSubRepoPermsStore{
-		DeleteByUserFunc: &SubRepoPermsStoreDeleteByUserFunc{
-			defaultHook: func(context.Context, int32) (r0 error) {
-				return
-			},
-		},
-		DoneFunc: &SubRepoPermsStoreDoneFunc{
-			defaultHook: func(error) (r0 error) {
-				return
-			},
-		},
-		GetFunc: &SubRepoPermsStoreGetFunc{
-			defaultHook: func(context.Context, int32, api.RepoID) (r0 *authz.SubRepoPermissions, r1 error) {
-				return
-			},
-		},
-		GetByUserFunc: &SubRepoPermsStoreGetByUserFunc{
-			defaultHook: func(context.Context, int32) (r0 map[api.RepoName]authz.SubRepoPermissions, r1 error) {
-				return
-			},
-		},
-		GetByUserAndServiceFunc: &SubRepoPermsStoreGetByUserAndServiceFunc{
-			defaultHook: func(context.Context, int32, string, string) (r0 map[api.ExternalRepoSpec]authz.SubRepoPermissions, r1 error) {
-				return
-			},
-		},
-		HandleFunc: &SubRepoPermsStoreHandleFunc{
-			defaultHook: func() (r0 basestore.TransactableHandle) {
-				return
-			},
-		},
-		RepoIDSupportedFunc: &SubRepoPermsStoreRepoIDSupportedFunc{
-			defaultHook: func(context.Context, api.RepoID) (r0 bool, r1 error) {
-				return
-			},
-		},
-		RepoSupportedFunc: &SubRepoPermsStoreRepoSupportedFunc{
-			defaultHook: func(context.Context, api.RepoName) (r0 bool, r1 error) {
-				return
-			},
-		},
-		TransactFunc: &SubRepoPermsStoreTransactFunc{
-			defaultHook: func(context.Context) (r0 SubRepoPermsStore, r1 error) {
-				return
-			},
-		},
-		UpsertFunc: &SubRepoPermsStoreUpsertFunc{
-			defaultHook: func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) (r0 error) {
-				return
-			},
-		},
-		UpsertWithSpecFunc: &SubRepoPermsStoreUpsertWithSpecFunc{
-			defaultHook: func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) (r0 error) {
-				return
-			},
-		},
-		WithFunc: &SubRepoPermsStoreWithFunc{
-			defaultHook: func(basestore.ShareableStore) (r0 SubRepoPermsStore) {
-				return
-			},
-		},
-	}
-}
-
-// NewStrictMockSubRepoPermsStore creates a new mock of the
-// SubRepoPermsStore interface. All methods panic on invocation, unless
-// overwritten.
-func NewStrictMockSubRepoPermsStore() *MockSubRepoPermsStore {
-	return &MockSubRepoPermsStore{
-		DeleteByUserFunc: &SubRepoPermsStoreDeleteByUserFunc{
-			defaultHook: func(context.Context, int32) error {
-				panic("unexpected invocation of MockSubRepoPermsStore.DeleteByUser")
-			},
-		},
-		DoneFunc: &SubRepoPermsStoreDoneFunc{
-			defaultHook: func(error) error {
-				panic("unexpected invocation of MockSubRepoPermsStore.Done")
-			},
-		},
-		GetFunc: &SubRepoPermsStoreGetFunc{
-			defaultHook: func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.Get")
-			},
-		},
-		GetByUserFunc: &SubRepoPermsStoreGetByUserFunc{
-			defaultHook: func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.GetByUser")
-			},
-		},
-		GetByUserAndServiceFunc: &SubRepoPermsStoreGetByUserAndServiceFunc{
-			defaultHook: func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.GetByUserAndService")
-			},
-		},
-		HandleFunc: &SubRepoPermsStoreHandleFunc{
-			defaultHook: func() basestore.TransactableHandle {
-				panic("unexpected invocation of MockSubRepoPermsStore.Handle")
-			},
-		},
-		RepoIDSupportedFunc: &SubRepoPermsStoreRepoIDSupportedFunc{
-			defaultHook: func(context.Context, api.RepoID) (bool, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.RepoIDSupported")
-			},
-		},
-		RepoSupportedFunc: &SubRepoPermsStoreRepoSupportedFunc{
-			defaultHook: func(context.Context, api.RepoName) (bool, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.RepoSupported")
-			},
-		},
-		TransactFunc: &SubRepoPermsStoreTransactFunc{
-			defaultHook: func(context.Context) (SubRepoPermsStore, error) {
-				panic("unexpected invocation of MockSubRepoPermsStore.Transact")
-			},
-		},
-		UpsertFunc: &SubRepoPermsStoreUpsertFunc{
-			defaultHook: func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error {
-				panic("unexpected invocation of MockSubRepoPermsStore.Upsert")
-			},
-		},
-		UpsertWithSpecFunc: &SubRepoPermsStoreUpsertWithSpecFunc{
-			defaultHook: func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error {
-				panic("unexpected invocation of MockSubRepoPermsStore.UpsertWithSpec")
-			},
-		},
-		WithFunc: &SubRepoPermsStoreWithFunc{
-			defaultHook: func(basestore.ShareableStore) SubRepoPermsStore {
-				panic("unexpected invocation of MockSubRepoPermsStore.With")
-			},
-		},
-	}
-}
-
-// NewMockSubRepoPermsStoreFrom creates a new mock of the
-// MockSubRepoPermsStore interface. All methods delegate to the given
-// implementation, unless overwritten.
-func NewMockSubRepoPermsStoreFrom(i SubRepoPermsStore) *MockSubRepoPermsStore {
-	return &MockSubRepoPermsStore{
-		DeleteByUserFunc: &SubRepoPermsStoreDeleteByUserFunc{
-			defaultHook: i.DeleteByUser,
-		},
-		DoneFunc: &SubRepoPermsStoreDoneFunc{
-			defaultHook: i.Done,
-		},
-		GetFunc: &SubRepoPermsStoreGetFunc{
-			defaultHook: i.Get,
-		},
-		GetByUserFunc: &SubRepoPermsStoreGetByUserFunc{
-			defaultHook: i.GetByUser,
-		},
-		GetByUserAndServiceFunc: &SubRepoPermsStoreGetByUserAndServiceFunc{
-			defaultHook: i.GetByUserAndService,
-		},
-		HandleFunc: &SubRepoPermsStoreHandleFunc{
-			defaultHook: i.Handle,
-		},
-		RepoIDSupportedFunc: &SubRepoPermsStoreRepoIDSupportedFunc{
-			defaultHook: i.RepoIDSupported,
-		},
-		RepoSupportedFunc: &SubRepoPermsStoreRepoSupportedFunc{
-			defaultHook: i.RepoSupported,
-		},
-		TransactFunc: &SubRepoPermsStoreTransactFunc{
-			defaultHook: i.Transact,
-		},
-		UpsertFunc: &SubRepoPermsStoreUpsertFunc{
-			defaultHook: i.Upsert,
-		},
-		UpsertWithSpecFunc: &SubRepoPermsStoreUpsertWithSpecFunc{
-			defaultHook: i.UpsertWithSpec,
-		},
-		WithFunc: &SubRepoPermsStoreWithFunc{
-			defaultHook: i.With,
-		},
-	}
-}
-
-// SubRepoPermsStoreDeleteByUserFunc describes the behavior when the
-// DeleteByUser method of the parent MockSubRepoPermsStore instance is
-// invoked.
-type SubRepoPermsStoreDeleteByUserFunc struct {
-	defaultHook func(context.Context, int32) error
-	hooks       []func(context.Context, int32) error
-	history     []SubRepoPermsStoreDeleteByUserFuncCall
-	mutex       sync.Mutex
-}
-
-// DeleteByUser delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) DeleteByUser(v0 context.Context, v1 int32) error {
-	r0 := m.DeleteByUserFunc.nextHook()(v0, v1)
-	m.DeleteByUserFunc.appendCall(SubRepoPermsStoreDeleteByUserFuncCall{v0, v1, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the DeleteByUser method
-// of the parent MockSubRepoPermsStore instance is invoked and the hook
-// queue is empty.
-func (f *SubRepoPermsStoreDeleteByUserFunc) SetDefaultHook(hook func(context.Context, int32) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// DeleteByUser method of the parent MockSubRepoPermsStore instance invokes
-// the hook at the front of the queue and discards it. After the queue is
-// empty, the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreDeleteByUserFunc) PushHook(hook func(context.Context, int32) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreDeleteByUserFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, int32) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreDeleteByUserFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, int32) error {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreDeleteByUserFunc) nextHook() func(context.Context, int32) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreDeleteByUserFunc) appendCall(r0 SubRepoPermsStoreDeleteByUserFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreDeleteByUserFuncCall
-// objects describing the invocations of this function.
-func (f *SubRepoPermsStoreDeleteByUserFunc) History() []SubRepoPermsStoreDeleteByUserFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreDeleteByUserFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreDeleteByUserFuncCall is an object that describes an
-// invocation of method DeleteByUser on an instance of
-// MockSubRepoPermsStore.
-type SubRepoPermsStoreDeleteByUserFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreDeleteByUserFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreDeleteByUserFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// SubRepoPermsStoreDoneFunc describes the behavior when the Done method of
-// the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreDoneFunc struct {
-	defaultHook func(error) error
-	hooks       []func(error) error
-	history     []SubRepoPermsStoreDoneFuncCall
-	mutex       sync.Mutex
-}
-
-// Done delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) Done(v0 error) error {
-	r0 := m.DoneFunc.nextHook()(v0)
-	m.DoneFunc.appendCall(SubRepoPermsStoreDoneFuncCall{v0, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the Done method of the
-// parent MockSubRepoPermsStore instance is invoked and the hook queue is
-// empty.
-func (f *SubRepoPermsStoreDoneFunc) SetDefaultHook(hook func(error) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Done method of the parent MockSubRepoPermsStore instance invokes the hook
-// at the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreDoneFunc) PushHook(hook func(error) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreDoneFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(error) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreDoneFunc) PushReturn(r0 error) {
-	f.PushHook(func(error) error {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreDoneFunc) nextHook() func(error) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreDoneFunc) appendCall(r0 SubRepoPermsStoreDoneFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreDoneFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreDoneFunc) History() []SubRepoPermsStoreDoneFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreDoneFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreDoneFuncCall is an object that describes an invocation
-// of method Done on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreDoneFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 error
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreDoneFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreDoneFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// SubRepoPermsStoreGetFunc describes the behavior when the Get method of
-// the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreGetFunc struct {
-	defaultHook func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error)
-	hooks       []func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error)
-	history     []SubRepoPermsStoreGetFuncCall
-	mutex       sync.Mutex
-}
-
-// Get delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) Get(v0 context.Context, v1 int32, v2 api.RepoID) (*authz.SubRepoPermissions, error) {
-	r0, r1 := m.GetFunc.nextHook()(v0, v1, v2)
-	m.GetFunc.appendCall(SubRepoPermsStoreGetFuncCall{v0, v1, v2, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the Get method of the
-// parent MockSubRepoPermsStore instance is invoked and the hook queue is
-// empty.
-func (f *SubRepoPermsStoreGetFunc) SetDefaultHook(hook func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Get method of the parent MockSubRepoPermsStore instance invokes the hook
-// at the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreGetFunc) PushHook(hook func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreGetFunc) SetDefaultReturn(r0 *authz.SubRepoPermissions, r1 error) {
-	f.SetDefaultHook(func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreGetFunc) PushReturn(r0 *authz.SubRepoPermissions, r1 error) {
-	f.PushHook(func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreGetFunc) nextHook() func(context.Context, int32, api.RepoID) (*authz.SubRepoPermissions, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreGetFunc) appendCall(r0 SubRepoPermsStoreGetFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreGetFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreGetFunc) History() []SubRepoPermsStoreGetFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreGetFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreGetFuncCall is an object that describes an invocation of
-// method Get on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreGetFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 api.RepoID
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 *authz.SubRepoPermissions
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreGetFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreGetFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreGetByUserFunc describes the behavior when the GetByUser
-// method of the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreGetByUserFunc struct {
-	defaultHook func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error)
-	hooks       []func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error)
-	history     []SubRepoPermsStoreGetByUserFuncCall
-	mutex       sync.Mutex
-}
-
-// GetByUser delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) GetByUser(v0 context.Context, v1 int32) (map[api.RepoName]authz.SubRepoPermissions, error) {
-	r0, r1 := m.GetByUserFunc.nextHook()(v0, v1)
-	m.GetByUserFunc.appendCall(SubRepoPermsStoreGetByUserFuncCall{v0, v1, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the GetByUser method of
-// the parent MockSubRepoPermsStore instance is invoked and the hook queue
-// is empty.
-func (f *SubRepoPermsStoreGetByUserFunc) SetDefaultHook(hook func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// GetByUser method of the parent MockSubRepoPermsStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreGetByUserFunc) PushHook(hook func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreGetByUserFunc) SetDefaultReturn(r0 map[api.RepoName]authz.SubRepoPermissions, r1 error) {
-	f.SetDefaultHook(func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreGetByUserFunc) PushReturn(r0 map[api.RepoName]authz.SubRepoPermissions, r1 error) {
-	f.PushHook(func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreGetByUserFunc) nextHook() func(context.Context, int32) (map[api.RepoName]authz.SubRepoPermissions, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreGetByUserFunc) appendCall(r0 SubRepoPermsStoreGetByUserFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreGetByUserFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreGetByUserFunc) History() []SubRepoPermsStoreGetByUserFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreGetByUserFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreGetByUserFuncCall is an object that describes an
-// invocation of method GetByUser on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreGetByUserFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 map[api.RepoName]authz.SubRepoPermissions
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreGetByUserFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreGetByUserFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreGetByUserAndServiceFunc describes the behavior when the
-// GetByUserAndService method of the parent MockSubRepoPermsStore instance
-// is invoked.
-type SubRepoPermsStoreGetByUserAndServiceFunc struct {
-	defaultHook func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error)
-	hooks       []func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error)
-	history     []SubRepoPermsStoreGetByUserAndServiceFuncCall
-	mutex       sync.Mutex
-}
-
-// GetByUserAndService delegates to the next hook function in the queue and
-// stores the parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) GetByUserAndService(v0 context.Context, v1 int32, v2 string, v3 string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error) {
-	r0, r1 := m.GetByUserAndServiceFunc.nextHook()(v0, v1, v2, v3)
-	m.GetByUserAndServiceFunc.appendCall(SubRepoPermsStoreGetByUserAndServiceFuncCall{v0, v1, v2, v3, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the GetByUserAndService
-// method of the parent MockSubRepoPermsStore instance is invoked and the
-// hook queue is empty.
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) SetDefaultHook(hook func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// GetByUserAndService method of the parent MockSubRepoPermsStore instance
-// invokes the hook at the front of the queue and discards it. After the
-// queue is empty, the default hook function is invoked for any future
-// action.
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) PushHook(hook func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) SetDefaultReturn(r0 map[api.ExternalRepoSpec]authz.SubRepoPermissions, r1 error) {
-	f.SetDefaultHook(func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) PushReturn(r0 map[api.ExternalRepoSpec]authz.SubRepoPermissions, r1 error) {
-	f.PushHook(func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) nextHook() func(context.Context, int32, string, string) (map[api.ExternalRepoSpec]authz.SubRepoPermissions, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) appendCall(r0 SubRepoPermsStoreGetByUserAndServiceFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of
-// SubRepoPermsStoreGetByUserAndServiceFuncCall objects describing the
-// invocations of this function.
-func (f *SubRepoPermsStoreGetByUserAndServiceFunc) History() []SubRepoPermsStoreGetByUserAndServiceFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreGetByUserAndServiceFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreGetByUserAndServiceFuncCall is an object that describes
-// an invocation of method GetByUserAndService on an instance of
-// MockSubRepoPermsStore.
-type SubRepoPermsStoreGetByUserAndServiceFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 string
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 string
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 map[api.ExternalRepoSpec]authz.SubRepoPermissions
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreGetByUserAndServiceFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreGetByUserAndServiceFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreHandleFunc describes the behavior when the Handle method
-// of the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreHandleFunc struct {
-	defaultHook func() basestore.TransactableHandle
-	hooks       []func() basestore.TransactableHandle
-	history     []SubRepoPermsStoreHandleFuncCall
-	mutex       sync.Mutex
-}
-
-// Handle delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) Handle() basestore.TransactableHandle {
-	r0 := m.HandleFunc.nextHook()()
-	m.HandleFunc.appendCall(SubRepoPermsStoreHandleFuncCall{r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the Handle method of the
-// parent MockSubRepoPermsStore instance is invoked and the hook queue is
-// empty.
-func (f *SubRepoPermsStoreHandleFunc) SetDefaultHook(hook func() basestore.TransactableHandle) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Handle method of the parent MockSubRepoPermsStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreHandleFunc) PushHook(hook func() basestore.TransactableHandle) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreHandleFunc) SetDefaultReturn(r0 basestore.TransactableHandle) {
-	f.SetDefaultHook(func() basestore.TransactableHandle {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreHandleFunc) PushReturn(r0 basestore.TransactableHandle) {
-	f.PushHook(func() basestore.TransactableHandle {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreHandleFunc) nextHook() func() basestore.TransactableHandle {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreHandleFunc) appendCall(r0 SubRepoPermsStoreHandleFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreHandleFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreHandleFunc) History() []SubRepoPermsStoreHandleFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreHandleFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreHandleFuncCall is an object that describes an invocation
-// of method Handle on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreHandleFuncCall struct {
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 basestore.TransactableHandle
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreHandleFuncCall) Args() []interface{} {
-	return []interface{}{}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreHandleFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// SubRepoPermsStoreRepoIDSupportedFunc describes the behavior when the
-// RepoIDSupported method of the parent MockSubRepoPermsStore instance is
-// invoked.
-type SubRepoPermsStoreRepoIDSupportedFunc struct {
-	defaultHook func(context.Context, api.RepoID) (bool, error)
-	hooks       []func(context.Context, api.RepoID) (bool, error)
-	history     []SubRepoPermsStoreRepoIDSupportedFuncCall
-	mutex       sync.Mutex
-}
-
-// RepoIDSupported delegates to the next hook function in the queue and
-// stores the parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) RepoIDSupported(v0 context.Context, v1 api.RepoID) (bool, error) {
-	r0, r1 := m.RepoIDSupportedFunc.nextHook()(v0, v1)
-	m.RepoIDSupportedFunc.appendCall(SubRepoPermsStoreRepoIDSupportedFuncCall{v0, v1, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the RepoIDSupported
-// method of the parent MockSubRepoPermsStore instance is invoked and the
-// hook queue is empty.
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) SetDefaultHook(hook func(context.Context, api.RepoID) (bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// RepoIDSupported method of the parent MockSubRepoPermsStore instance
-// invokes the hook at the front of the queue and discards it. After the
-// queue is empty, the default hook function is invoked for any future
-// action.
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) PushHook(hook func(context.Context, api.RepoID) (bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) SetDefaultReturn(r0 bool, r1 error) {
-	f.SetDefaultHook(func(context.Context, api.RepoID) (bool, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) PushReturn(r0 bool, r1 error) {
-	f.PushHook(func(context.Context, api.RepoID) (bool, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) nextHook() func(context.Context, api.RepoID) (bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) appendCall(r0 SubRepoPermsStoreRepoIDSupportedFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreRepoIDSupportedFuncCall
-// objects describing the invocations of this function.
-func (f *SubRepoPermsStoreRepoIDSupportedFunc) History() []SubRepoPermsStoreRepoIDSupportedFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreRepoIDSupportedFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreRepoIDSupportedFuncCall is an object that describes an
-// invocation of method RepoIDSupported on an instance of
-// MockSubRepoPermsStore.
-type SubRepoPermsStoreRepoIDSupportedFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 api.RepoID
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 bool
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreRepoIDSupportedFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreRepoIDSupportedFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreRepoSupportedFunc describes the behavior when the
-// RepoSupported method of the parent MockSubRepoPermsStore instance is
-// invoked.
-type SubRepoPermsStoreRepoSupportedFunc struct {
-	defaultHook func(context.Context, api.RepoName) (bool, error)
-	hooks       []func(context.Context, api.RepoName) (bool, error)
-	history     []SubRepoPermsStoreRepoSupportedFuncCall
-	mutex       sync.Mutex
-}
-
-// RepoSupported delegates to the next hook function in the queue and stores
-// the parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) RepoSupported(v0 context.Context, v1 api.RepoName) (bool, error) {
-	r0, r1 := m.RepoSupportedFunc.nextHook()(v0, v1)
-	m.RepoSupportedFunc.appendCall(SubRepoPermsStoreRepoSupportedFuncCall{v0, v1, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the RepoSupported method
-// of the parent MockSubRepoPermsStore instance is invoked and the hook
-// queue is empty.
-func (f *SubRepoPermsStoreRepoSupportedFunc) SetDefaultHook(hook func(context.Context, api.RepoName) (bool, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// RepoSupported method of the parent MockSubRepoPermsStore instance invokes
-// the hook at the front of the queue and discards it. After the queue is
-// empty, the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreRepoSupportedFunc) PushHook(hook func(context.Context, api.RepoName) (bool, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreRepoSupportedFunc) SetDefaultReturn(r0 bool, r1 error) {
-	f.SetDefaultHook(func(context.Context, api.RepoName) (bool, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreRepoSupportedFunc) PushReturn(r0 bool, r1 error) {
-	f.PushHook(func(context.Context, api.RepoName) (bool, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreRepoSupportedFunc) nextHook() func(context.Context, api.RepoName) (bool, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreRepoSupportedFunc) appendCall(r0 SubRepoPermsStoreRepoSupportedFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreRepoSupportedFuncCall
-// objects describing the invocations of this function.
-func (f *SubRepoPermsStoreRepoSupportedFunc) History() []SubRepoPermsStoreRepoSupportedFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreRepoSupportedFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreRepoSupportedFuncCall is an object that describes an
-// invocation of method RepoSupported on an instance of
-// MockSubRepoPermsStore.
-type SubRepoPermsStoreRepoSupportedFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 api.RepoName
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 bool
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreRepoSupportedFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreRepoSupportedFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreTransactFunc describes the behavior when the Transact
-// method of the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreTransactFunc struct {
-	defaultHook func(context.Context) (SubRepoPermsStore, error)
-	hooks       []func(context.Context) (SubRepoPermsStore, error)
-	history     []SubRepoPermsStoreTransactFuncCall
-	mutex       sync.Mutex
-}
-
-// Transact delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) Transact(v0 context.Context) (SubRepoPermsStore, error) {
-	r0, r1 := m.TransactFunc.nextHook()(v0)
-	m.TransactFunc.appendCall(SubRepoPermsStoreTransactFuncCall{v0, r0, r1})
-	return r0, r1
-}
-
-// SetDefaultHook sets function that is called when the Transact method of
-// the parent MockSubRepoPermsStore instance is invoked and the hook queue
-// is empty.
-func (f *SubRepoPermsStoreTransactFunc) SetDefaultHook(hook func(context.Context) (SubRepoPermsStore, error)) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Transact method of the parent MockSubRepoPermsStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreTransactFunc) PushHook(hook func(context.Context) (SubRepoPermsStore, error)) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreTransactFunc) SetDefaultReturn(r0 SubRepoPermsStore, r1 error) {
-	f.SetDefaultHook(func(context.Context) (SubRepoPermsStore, error) {
-		return r0, r1
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreTransactFunc) PushReturn(r0 SubRepoPermsStore, r1 error) {
-	f.PushHook(func(context.Context) (SubRepoPermsStore, error) {
-		return r0, r1
-	})
-}
-
-func (f *SubRepoPermsStoreTransactFunc) nextHook() func(context.Context) (SubRepoPermsStore, error) {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreTransactFunc) appendCall(r0 SubRepoPermsStoreTransactFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreTransactFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreTransactFunc) History() []SubRepoPermsStoreTransactFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreTransactFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreTransactFuncCall is an object that describes an
-// invocation of method Transact on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreTransactFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 SubRepoPermsStore
-	// Result1 is the value of the 2nd result returned from this method
-	// invocation.
-	Result1 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreTransactFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreTransactFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0, c.Result1}
-}
-
-// SubRepoPermsStoreUpsertFunc describes the behavior when the Upsert method
-// of the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreUpsertFunc struct {
-	defaultHook func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error
-	hooks       []func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error
-	history     []SubRepoPermsStoreUpsertFuncCall
-	mutex       sync.Mutex
-}
-
-// Upsert delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) Upsert(v0 context.Context, v1 int32, v2 api.RepoID, v3 authz.SubRepoPermissions) error {
-	r0 := m.UpsertFunc.nextHook()(v0, v1, v2, v3)
-	m.UpsertFunc.appendCall(SubRepoPermsStoreUpsertFuncCall{v0, v1, v2, v3, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the Upsert method of the
-// parent MockSubRepoPermsStore instance is invoked and the hook queue is
-// empty.
-func (f *SubRepoPermsStoreUpsertFunc) SetDefaultHook(hook func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// Upsert method of the parent MockSubRepoPermsStore instance invokes the
-// hook at the front of the queue and discards it. After the queue is empty,
-// the default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreUpsertFunc) PushHook(hook func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreUpsertFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreUpsertFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreUpsertFunc) nextHook() func(context.Context, int32, api.RepoID, authz.SubRepoPermissions) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreUpsertFunc) appendCall(r0 SubRepoPermsStoreUpsertFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreUpsertFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreUpsertFunc) History() []SubRepoPermsStoreUpsertFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreUpsertFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreUpsertFuncCall is an object that describes an invocation
-// of method Upsert on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreUpsertFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 api.RepoID
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 authz.SubRepoPermissions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreUpsertFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreUpsertFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// SubRepoPermsStoreUpsertWithSpecFunc describes the behavior when the
-// UpsertWithSpec method of the parent MockSubRepoPermsStore instance is
-// invoked.
-type SubRepoPermsStoreUpsertWithSpecFunc struct {
-	defaultHook func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error
-	hooks       []func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error
-	history     []SubRepoPermsStoreUpsertWithSpecFuncCall
-	mutex       sync.Mutex
-}
-
-// UpsertWithSpec delegates to the next hook function in the queue and
-// stores the parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) UpsertWithSpec(v0 context.Context, v1 int32, v2 api.ExternalRepoSpec, v3 authz.SubRepoPermissions) error {
-	r0 := m.UpsertWithSpecFunc.nextHook()(v0, v1, v2, v3)
-	m.UpsertWithSpecFunc.appendCall(SubRepoPermsStoreUpsertWithSpecFuncCall{v0, v1, v2, v3, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the UpsertWithSpec
-// method of the parent MockSubRepoPermsStore instance is invoked and the
-// hook queue is empty.
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) SetDefaultHook(hook func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// UpsertWithSpec method of the parent MockSubRepoPermsStore instance
-// invokes the hook at the front of the queue and discards it. After the
-// queue is empty, the default hook function is invoked for any future
-// action.
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) PushHook(hook func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) SetDefaultReturn(r0 error) {
-	f.SetDefaultHook(func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) PushReturn(r0 error) {
-	f.PushHook(func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) nextHook() func(context.Context, int32, api.ExternalRepoSpec, authz.SubRepoPermissions) error {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) appendCall(r0 SubRepoPermsStoreUpsertWithSpecFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreUpsertWithSpecFuncCall
-// objects describing the invocations of this function.
-func (f *SubRepoPermsStoreUpsertWithSpecFunc) History() []SubRepoPermsStoreUpsertWithSpecFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreUpsertWithSpecFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreUpsertWithSpecFuncCall is an object that describes an
-// invocation of method UpsertWithSpec on an instance of
-// MockSubRepoPermsStore.
-type SubRepoPermsStoreUpsertWithSpecFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 context.Context
-	// Arg1 is the value of the 2nd argument passed to this method
-	// invocation.
-	Arg1 int32
-	// Arg2 is the value of the 3rd argument passed to this method
-	// invocation.
-	Arg2 api.ExternalRepoSpec
-	// Arg3 is the value of the 4th argument passed to this method
-	// invocation.
-	Arg3 authz.SubRepoPermissions
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 error
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreUpsertWithSpecFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0, c.Arg1, c.Arg2, c.Arg3}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreUpsertWithSpecFuncCall) Results() []interface{} {
-	return []interface{}{c.Result0}
-}
-
-// SubRepoPermsStoreWithFunc describes the behavior when the With method of
-// the parent MockSubRepoPermsStore instance is invoked.
-type SubRepoPermsStoreWithFunc struct {
-	defaultHook func(basestore.ShareableStore) SubRepoPermsStore
-	hooks       []func(basestore.ShareableStore) SubRepoPermsStore
-	history     []SubRepoPermsStoreWithFuncCall
-	mutex       sync.Mutex
-}
-
-// With delegates to the next hook function in the queue and stores the
-// parameter and result values of this invocation.
-func (m *MockSubRepoPermsStore) With(v0 basestore.ShareableStore) SubRepoPermsStore {
-	r0 := m.WithFunc.nextHook()(v0)
-	m.WithFunc.appendCall(SubRepoPermsStoreWithFuncCall{v0, r0})
-	return r0
-}
-
-// SetDefaultHook sets function that is called when the With method of the
-// parent MockSubRepoPermsStore instance is invoked and the hook queue is
-// empty.
-func (f *SubRepoPermsStoreWithFunc) SetDefaultHook(hook func(basestore.ShareableStore) SubRepoPermsStore) {
-	f.defaultHook = hook
-}
-
-// PushHook adds a function to the end of hook queue. Each invocation of the
-// With method of the parent MockSubRepoPermsStore instance invokes the hook
-// at the front of the queue and discards it. After the queue is empty, the
-// default hook function is invoked for any future action.
-func (f *SubRepoPermsStoreWithFunc) PushHook(hook func(basestore.ShareableStore) SubRepoPermsStore) {
-	f.mutex.Lock()
-	f.hooks = append(f.hooks, hook)
-	f.mutex.Unlock()
-}
-
-// SetDefaultReturn calls SetDefaultHook with a function that returns the
-// given values.
-func (f *SubRepoPermsStoreWithFunc) SetDefaultReturn(r0 SubRepoPermsStore) {
-	f.SetDefaultHook(func(basestore.ShareableStore) SubRepoPermsStore {
-		return r0
-	})
-}
-
-// PushReturn calls PushHook with a function that returns the given values.
-func (f *SubRepoPermsStoreWithFunc) PushReturn(r0 SubRepoPermsStore) {
-	f.PushHook(func(basestore.ShareableStore) SubRepoPermsStore {
-		return r0
-	})
-}
-
-func (f *SubRepoPermsStoreWithFunc) nextHook() func(basestore.ShareableStore) SubRepoPermsStore {
-	f.mutex.Lock()
-	defer f.mutex.Unlock()
-
-	if len(f.hooks) == 0 {
-		return f.defaultHook
-	}
-
-	hook := f.hooks[0]
-	f.hooks = f.hooks[1:]
-	return hook
-}
-
-func (f *SubRepoPermsStoreWithFunc) appendCall(r0 SubRepoPermsStoreWithFuncCall) {
-	f.mutex.Lock()
-	f.history = append(f.history, r0)
-	f.mutex.Unlock()
-}
-
-// History returns a sequence of SubRepoPermsStoreWithFuncCall objects
-// describing the invocations of this function.
-func (f *SubRepoPermsStoreWithFunc) History() []SubRepoPermsStoreWithFuncCall {
-	f.mutex.Lock()
-	history := make([]SubRepoPermsStoreWithFuncCall, len(f.history))
-	copy(history, f.history)
-	f.mutex.Unlock()
-
-	return history
-}
-
-// SubRepoPermsStoreWithFuncCall is an object that describes an invocation
-// of method With on an instance of MockSubRepoPermsStore.
-type SubRepoPermsStoreWithFuncCall struct {
-	// Arg0 is the value of the 1st argument passed to this method
-	// invocation.
-	Arg0 basestore.ShareableStore
-	// Result0 is the value of the 1st result returned from this method
-	// invocation.
-	Result0 SubRepoPermsStore
-}
-
-// Args returns an interface slice containing the arguments of this
-// invocation.
-func (c SubRepoPermsStoreWithFuncCall) Args() []interface{} {
-	return []interface{}{c.Arg0}
-}
-
-// Results returns an interface slice containing the results of this
-// invocation.
-func (c SubRepoPermsStoreWithFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
 }
